propagating syntax modifications. Code compiles.

+++ testSuite/unit/MsqVertexTest.cpp
@@ -8,7 +8,7 @@
 //    E-MAIL: tleurent@mcs.anl.gov
 //
 // ORIG-DATE: 12-Nov-02 at 18:05:56
-//  LAST-MOD:  5-Dec-02 at 10:39:44 by Thomas Leurent
+//  LAST-MOD: 13-Dec-02 at 08:27:21 by Thomas Leurent
 //
 // DESCRIPTION:
 // ============
@@ -106,7 +106,7 @@ public:
     MSQ_CHKERR(err);
     
     // patch has only one element: an ideal hex.
-    int indices[8];
+    size_t indices[8];
     indices[0] = 0; indices[1] = 1; indices[2] = 2; indices[3] = 3;
     indices[4] = 4; indices[5] = 5; indices[6] = 6; indices[7] = 7;
     one_hex_patch.add_element(NULL, NULL, indices, HEXAHEDRON, err);
@@ -138,7 +138,7 @@ public:
     MSQ_CHKERR(err);
     
       // patch has only one element: an ideal tet.
-    int indices_tet[4];
+    size_t indices_tet[4];
     indices_tet[0] = 0; indices_tet[1] = 1; indices_tet[2] = 2;
     indices_tet[3] = 3;
     one_tet_patch.add_element(NULL, NULL, indices_tet, TETRAHEDRON, err);
@@ -169,7 +169,7 @@ public:
     MSQ_CHKERR(err);
     
       // patch has only one element: an ideal quad.
-    int indices_qua[4];
+    size_t indices_qua[4];
     indices_qua[0] = 0; indices_qua[1] = 1; indices_qua[2] = 2;
     indices_qua[3] = 3;
     one_qua_patch.add_element(NULL, NULL, indices_qua, QUADRILATERAL, err);
@@ -196,7 +196,7 @@ public:
     MSQ_CHKERR(err);
     
       // patch has only one element: an ideal tri
-    int indices_tri[3];
+    size_t indices_tri[3];
     indices_tri[0] = 0; indices_tri[1] = 1; indices_tri[2] = 2;
     one_tri_patch.add_element(NULL, NULL, indices_tri, TRIANGLE, err);
     MSQ_CHKERR(err);
